requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.6
django-dotenv==1.4.2
django-filter==23.5
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==24.0
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
